+++ tests/framework/db/ibm/IbmActiveRecordTest.php
@@ -1,0 +1,145 @@
+<?php
+
+namespace yiiunit\framework\db\ibm;
+
+use yiiunit\data\ar\NullValues;
+use yiiunit\data\ar\Type;
+use yiiunit\framework\db\ActiveRecordTest;
+
+/**
+ * @group db
+ * @group ibm
+ */
+class IbmActiveRecordTest extends ActiveRecordTest
+{
+    public $driverName = 'ibm';
+
+    public function testFindAsArray()
+    {
+        /* @var $customerClass \yii\db\ActiveRecordInterface */
+        $customerClass = $this->getCustomerClass();
+
+        // asArray
+        $customer = $customerClass::find()->where(['id' => 2])->asArray()->one();
+        // clob: segmentation fault error
+        unset($customer['address']);
+        $this->assertEquals([
+            'id' => '2',
+            'email' => 'user2@example.com',
+            'name' => 'user2',
+            'status' => '1',
+            'profile_id' => null,
+        ], $customer);
+
+        // find all asArray
+        $customers = $customerClass::find()->asArray()->all();
+        $this->assertEquals(3, count($customers));
+        $this->assertArrayHasKey('id', $customers[0]);
+        $this->assertArrayHasKey('name', $customers[0]);
+        $this->assertArrayHasKey('email', $customers[0]);
+        $this->assertArrayHasKey('address', $customers[0]);
+        $this->assertArrayHasKey('status', $customers[0]);
+        $this->assertArrayHasKey('id', $customers[1]);
+        $this->assertArrayHasKey('name', $customers[1]);
+        $this->assertArrayHasKey('email', $customers[1]);
+        $this->assertArrayHasKey('address', $customers[1]);
+        $this->assertArrayHasKey('status', $customers[1]);
+        $this->assertArrayHasKey('id', $customers[2]);
+        $this->assertArrayHasKey('name', $customers[2]);
+        $this->assertArrayHasKey('email', $customers[2]);
+        $this->assertArrayHasKey('address', $customers[2]);
+        $this->assertArrayHasKey('status', $customers[2]);
+    }
+
+    public function testStoreNull()
+    {
+        $record = new NullValues();
+        $this->assertNull($record->var1);
+        $this->assertNull($record->var2);
+        $this->assertNull($record->var3);
+        $this->assertNull($record->stringcol);
+
+        $record->var1 = 123;
+        $record->var2 = 456;
+        $record->var3 = 789;
+        $record->stringcol = 'hello!';
+
+        $record->save(false);
+        $this->assertTrue($record->refresh());
+
+        $this->assertEquals(123, $record->var1);
+        $this->assertEquals(456, $record->var2);
+        $this->assertEquals(789, $record->var3);
+        $this->assertEquals('hello!', $record->stringcol);
+
+        $record->var1 = null;
+        $record->var2 = null;
+        $record->var3 = null;
+        $record->stringcol = null;
+
+        $record->save(false);
+        $this->assertTrue($record->refresh());
+
+        $this->assertNull($record->var1);
+        $this->assertNull($record->var2);
+        $this->assertNull($record->var3);
+        $this->assertNull($record->stringcol);
+
+        $record->var1 = 0;
+        $record->var2 = 0;
+        $record->var3 = 0;
+        $record->stringcol = '';
+
+        $record->save(false);
+        $this->assertTrue($record->refresh());
+
+        $this->assertEquals(0, $record->var1);
+        $this->assertEquals(0, $record->var2);
+        $this->assertEquals(0, $record->var3);
+        $this->assertEquals('', $record->stringcol);
+    }
+
+    public function testStoreEmpty()
+    {
+        $record = new NullValues();
+
+        // this is to simulate empty html form submission
+        $record->var1 = '';
+        $record->var2 = '';
+        $record->var3 = '';
+        $record->stringcol = '';
+
+        $record->save(false);
+        $this->assertTrue($record->refresh());
+
+        // https://github.com/yiisoft/yii2/commit/34945b0b69011bc7cab684c7f7095d837892a0d4#commitcomment-4458225
+        $this->assertTrue($record->var1 === $record->var2);
+        $this->assertTrue($record->var2 === $record->var3);
+    }
+
+    public function testCastValues()
+    {
+        $model = new Type();
+        $model->int_col = 123;
+        $model->int_col2 = 456;
+        $model->smallint_col = 42;
+        $model->char_col = '1337';
+        $model->char_col2 = 'test';
+        $model->char_col3 = 'test123';
+        $model->float_col = 3.742;
+        $model->float_col2 = 42.1337;
+        $model->bool_col = true;
+        $model->bool_col2 = false;
+        $model->save(false);
+
+        /* @var $model Type */
+        $model = Type::find()->one();
+        $this->assertSame(123, $model->int_col);
+        $this->assertSame(456, $model->int_col2);
+        $this->assertSame(42, $model->smallint_col);
+        $this->assertSame('1337', trim($model->char_col));
+        $this->assertSame('test', $model->char_col2);
+        // clob: segmentation fault error
+        // $this->assertSame('test123', $model->char_col3);
+    }
+}